Version_1.0/requirements.txt
pywin32
Kivy==2.3.0
bcrypt==4.2.0
reportlab==4.2.2
screeninfo==0.8.1
cryptography==43.0.1
pycryptodomex==3.20.0
pyinstaller
